Implemented Food Scriptable Object

+++ Assets/Scripts/Items/ItemsModel.cs
@@ -5,8 +5,8 @@ public class ItemsModel
 { 
     public int FoodSpawnIntervalDelay { get; private set; }
     public int PowerUpSpawnIntervalDelay { get; private set; }
-    public float SpawnAreaWidth { get; set; }
-    public float SpawnAreaHeight { get; set; } 
+    public int SpawnAreaWidth { get; set; }
+    public int SpawnAreaHeight { get; set; } 
     public bool GameRunning { get; set; } = true;
     public bool FoodEaten { get; set; } = false;
     public bool PowerUpEaten { get; set; } = false;
@@ -19,7 +19,7 @@ public class ItemsModel
 
         FoodSpawnIntervalDelay = itemsSO.foodSpawnIntervalDelay;
         PowerUpSpawnIntervalDelay = itemsSO.powerUpSpawnIntervalDelay;
-        SpawnAreaWidth = itemsSO.itemSpawnArea.x;
-        SpawnAreaHeight = itemsSO.itemSpawnArea.y;
+        SpawnAreaWidth = (int)itemsSO.itemSpawnArea.x;
+        SpawnAreaHeight = (int)itemsSO.itemSpawnArea.y;
     }
 }
